make warm-init use same build-order.lisp-expr as XC and cold-init
Prior to this patch, warm.lisp maintained its own list of files to
load during warm initialization.  I merged that list into
build-order.lisp-expr and added a :WARM option to indicate that a file
should be loaded during warm, rather than cold, initialization.  Warm
init does not load build-order.lisp-expr directly but instead reads a
file output by make-host-2.lisp with a list of files to compile.

Remaining kludges to fix (or not) are: (1) use logical pathnames
instead of unix-style pathnames; (2) reuse the build functions in
shared.lisp during warm init for compiling files.  Currently
shared.lisp is code only used during cold build--it is in the SB-COLD
package.  Some of that machinery would need to move to a warm-loaded
package (SB!IMPL?) for reuse in warm initialization.  If any more
experienced SBCL hackers have an idea for which package to place the
stem-processing code, I will refactor accordingly.

In the mean time, here is a an intermediary refactor that at least
centralizes the build order across all stages of the process.