moved declarations into one file loaded on both host and target, and some very ambitious and broken changes
I'm running into some problems due to DEFCLASS, a macro, requiring a
lot of functions available when cross-compiling.  I'm scared of the
prospects: how much of the system will need to be loaded on the host?

Hopefully we can avoid needing to compile things on the host that
conflict with ANSI CL standards (e.g. if we needed the PCL defclass
macro instead of the host's).  But I'm not confident