A bunch of changes to get PCL compiling as much as it needs to on host, and also cross-compiling without error. !cold-init error due to handling of built-in-classes
specifically things are breaking as a result of calling make-array in
the prototype forms defined in classes.lisp.  There is probably a
colder way to create prototype instances before the entire type system
is loaded, but at this point I'm not quite sure what braid.lisp does
with the built-in-classes table, and it's best to leave something
rather tricky like evading make-array until the braid process is
inspected further.

We will probably need to move the BIC-table and braid top-level code
into a !cold-init-pcl function of sorts that gets executed before
other top-level forms, right after other parts of the type system have
been initialized:

  (show-and-call !type-class-cold-init)
  (show-and-call !typedefs-cold-init)
  (show-and-call !world-lock-cold-init)
  (show-and-call !classes-cold-init)
  (show-and-call !early-type-cold-init)
  (show-and-call !late-type-cold-init)
  (show-and-call !alien-type-cold-init)
  (show-and-call !target-type-cold-init)
  (show-and-call !vm-type-cold-init)

But next on the agenda is debugging exactly where in the cold init
process this is even happening.

/can't /SHOW:
*READTABLE* isn't bound. (Try /SHOW0.)
/entering ERROR, argument list=..
0x04678897
/cold-printing ERROR arguments one by one..
FORMAT-CONTROL
SUBTYPEP dispatch for MAKE-ARRAY before the type system is ready
FORMAT-ARGUMENTS
NIL
/done cold-printing ERROR arguments
/entering INFINITE-ERROR-PROTECTOR, *CURRENT-ERROR-DEPTH*=..
0x00000000
Argh! error in cold init, halting
fatal error encountered in SBCL pid 16895(tid 140737353963264):
%PRIMITIVE HALT called; the party is over.

Welcome to LDB, a low-level debugger for the Lisp runtime environment.
ldb> backtrace
Backtrace:
   0: Foreign fp = 0x7ffff516a5c0, ra = 0x410b72
   1: Foreign fp = 0x7ffff516a5f0, ra = 0x413aa3
   2: Foreign fp = 0x7ffff516a630, ra = 0x413fda
   3: Foreign fp = 0x7ffff516aad0, ra = 0x7ffff79ca8f0
   4: (SB!C::TL-XEP COMMON-LISP::ERROR)
   5: (SB!C::TL-XEP SB!INT::BUG)
   6: SB!IMPL::%VECTOR-WIDETAG-AND-N-BITS
   7: (SB!C::TL-XEP COMMON-LISP::MAKE-ARRAY)
   8: SB!INT::SIMPLE-EVAL-IN-LEXENV
   9: (COMMON-LISP::LAMBDA (SB!PCL::KERNEL-BIC-ENTRY))
  10: (SB!C::TOP-LEVEL-FORM ())
  11: SB!KERNEL::!COLD-INIT